pyproject.toml
[tool.poetry]
name = "Spin_n_Stone"
version = "0.1.0"
description = ""
authors = ["shamblini <78273476+shamblini@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^3.2.3"
dj-database-url = "^1.0.0"
psycopg2-binary = "^2.9.5"
django-google-translate = "^1.1"
whitenoise = {extras = ["brotli"], version = "^6.2.0"}
gunicorn = "^20.1.0"
setuptools = "^65.5.0"
setproctitle = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
